Added Additional endpoints + Made it possible for bogus to create bookings that do not overlap, even on the same day. There is between 3 and 24 hours difference in the enddate of 1 booking and the start date of another booking

+++ KelBurgAPI/Controllers/RoomsController.cs
@@ -76,11 +76,11 @@ public class RoomsController : ControllerBase
         return Ok(rooms);
     }
 
-    [HttpPatch("changePriceId")]
-    public async Task<ActionResult<Rooms>> ChangePriceId(int roomIdToChange, int newPriceId)
+    [HttpPatch("changePrice")]
+    public async Task<ActionResult<Rooms>> ChangePriceId(int roomIdToChange, int newPrice)
     {
         Rooms roomTopatch = await _context.Rooms.FindAsync(roomIdToChange);
-        roomTopatch.PricePrNight = newPriceId;
+        roomTopatch.PricePrNight = newPrice;
         await _context.SaveChangesAsync();
         return Ok(roomTopatch);
     }
@@ -89,6 +89,12 @@ public class RoomsController : ControllerBase
     public async Task<ActionResult<Rooms>> DeleteRoom(int roomId)
     {
         Rooms room = await _context.Rooms.FindAsync(roomId);
+        
+        if (room == null)
+        {
+            return NotFound("Room not found");
+        }
+        
         _context.Rooms.Remove(room);
         await _context.SaveChangesAsync();
         return Ok(room);
